+++ src/index.jsx
@@ -635,14 +635,7 @@ FlipPage.propTypes = {
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node,
   ]),
-  orientation: (props, propName, componentName) => {
-    if (!/(vertical|horizontal)/.test(props[propName])) {
-      return new Error(`Invalid prop \`${propName}\` supplied to ` +
-        ` \`${componentName}\`. Expected \`horizontal\` or \`vertical\`. Validation failed.`);
-    }
-
-    return '';
-  },
+  orientation: PropTypes.oneOf(['vertical', 'horizontal']),
   animationDuration: PropTypes.number,
   treshold: PropTypes.number,
   maxAngle: PropTypes.number,